requirements.txt
# Core Django
Django==5.1.4
djangorestframework==3.15.2

# Django Extensions
django-cors-headers==4.5.0
django-filter==24.3
django-extensions==3.2.3
django-debug-toolbar==4.4.6
whitenoise==6.8.2

# Authentication & JWT
djangorestframework-simplejwt==5.3.0
django-allauth==65.3.0

# Database & Caching
psycopg2-binary==2.9.10
redis==5.2.1

# Async & Background Tasks
celery==5.4.0
django-celery-beat==2.7.0
channels==4.2.0
channels-redis==4.2.0

# Image Processing (Compatible with Python 3.13)
Pillow==11.0.0

# File Storage
boto3==1.35.84
django-storages==1.14.4

# Environment & Configuration
python-decouple==3.8

# Payment Processing
stripe==11.2.0

# Phone Number Handling
django-phonenumber-field==8.0.0
phonenumbers==8.13.51

# Development & Testing
pytest-django==4.9.0
factory-boy==3.3.1

# Production Server
gunicorn==23.0.0